feat: add custom ball rail bounce
One of which actually works, SimpleRailBounce. It uses normal
OnCollisionEnter-events. It's pretty basic and works the best, as long
as velocity is set directly, not added to the pre-existing.

TriggerRailBounce is an attempt to use "Is Trigger"-colliders, which
would avoid the normal collision behavior which is problematic when not
setting velocity directly. However, we then don't get the Collision data
and have to synthesize it ourselves, and that turned out to be a little
difficult - it's nearly working, but the box collider planes don't seem
to be mathed correctly.

RailBounce is an adaptation of the Unity example of
Physics.ContactEvent-subscription, which is a lower-level API. It kinda
worked, but all bounces are constant speed. That could probably be
overcome.

All the scripts seem to work best on box colliders, so there is a
manual approximation of the rails using boxes instead of a mesh. The
scripts need to be placed on those boxes. The physics layers should be
configured so that balls hit the box colliders and everything else hits
the mesh colliders.