feat: add spectator view & lobby manager
This adds a new scene, Spectator View, which can be built for a mobile
device. The script TargetDeviceChanger changes settings to fit
mobile/HMD depending on which scene is opened. Also added a GUI to
connect to specific addresses without recompiling.

The mobile target uses AR Foundation's ARCore plugin, which requires
OpenGL instead of Vulkan, and crashes if the Oculus plugin is loaded, so
some fancy switching is required.