feat: bloom & cue stick grab affordance
Turned on post-processing, added bloom with fix as described in
https://communityforums.atmeta.com/t5/Unity-VR-Development/Passthrough-and-post-processing/m-p/1131527/highlight/true#M23353
to fix pass-through being black. HDR also needed to be off in the render
pipeline. Tried to not apply bloom to UI using camera stacking, but
rendering an additional camera proved laggy on the Quest Pro.

Added affordance to grab cue stick with both hands.